% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{getData}
\alias{getData}
\title{Pull data for a site found with metScanR}
\usage{
getData(site_meta, start_date, end_date, temp_agg, token = NA)
}
\arguments{
\item{site_meta}{a metScanR list element.}

\item{start_date}{A YYYY-MM-DD hh:mm:ss formated start time}

\item{end_date}{A YYYY-MM-DD hh:mm:ss formated end time}

\item{token}{Optional, but required for Mesonet stations. See: \url{https://developers.synopticdata.com/mesonet/} for more information.}
}
\value{
Data frame data for the input site found with metScanR, if available.
}
\description{
This function takes an input of the site metadata from a metScanR search,
as well as start and end dates to download data for, and downloads
}
\examples{
\dontrun{
# Get Data for one SCAN site near CPER
cper_sites=metScanR::getNearby(siteID="NEON:CPER", radius = 30)
uscrn_out=metDownloadR::getData(site_meta = cper_sites$USW00094074,
                              start_date = "2018-10-01",
                             end_date = "2018-10-31", 
                             temp_agg="monthly")
# Get all october 2018 data from sites within a 5 km radius of CPER
out=lapply(
metScanR::getNearby(siteID="NEON:CPER", radius = 30),
metDownloadR::getData,
start_date = "2018-10-01",
end_date = "2018-10-31", 
temp_agg="daily")
}
}
\seealso{
Currently none
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
Josh Roberti\cr
}
\keyword{USCRN,}
\keyword{commissioning}
\keyword{data}
\keyword{data,}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
