% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uscrn_data.R
\name{getUSCRNData}
\alias{getUSCRNData}
\title{Pull data from meterologic stations in the USCRN network}
\usage{
getUSCRNData(temp_agg, sid, start_date, end_date)
}
\arguments{
\item{temp_agg}{The time scale of measurements to return.
Enter 'monthly', 'daily', 'hourly', or 'subhourly'.}

\item{sid}{The WBAN ID of a valid USCRN site.}

\item{start_date}{YYYY-MM-DD hh:mm:ss formated start time, in UTC.}

\item{end_date}{A YYYY-MM-DD hh:mm:ss formated end time, in UTC.}
}
\value{
Data frame of USCRN data.
}
\description{
Provide a valid GHCND or WBAN ID for a site in the USCRN, the desired
timescale of measurements, and requested begin and end times in POSIX format,
and a data frame will be returned to the global environment.
}
\examples{
\dontrun{
temp_agg <- "subhourly"
sid <- "USW00003047"
start_date <- as.POSIXct("2014-04-01 00:00:01", format="\%Y-\%m-\%d \%H:\%M:\%S", tz="UTC")
end_date <- as.POSIXct("2015-02-01 00:00:00", format = "\%Y-\%m-\%d \%H:\%M:\%S", tz="UTC")

grabUSCRN(temp_agg = temp_agg, TimeBgn = TimeBgn, TimeEnd = TimeEnd, sid = sid)
}
}
\seealso{
Currently none
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{USCRN,}
\keyword{commissioning}
\keyword{data}
\keyword{data,}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
