% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resTracking}
\alias{resTracking}
\title{Mapping of Temporal Resolution of Data by Platform and ID}
\format{Data frame}
\usage{
resTracking
}
\description{
A data frame of ID types, platforms, R packages to 
access sensor data, and the possible temporal resolutions of data 
served through each platform/ID type. Used to properly request 
data with \code{getData}.
}
\keyword{ACIS,}
\keyword{AL}
\keyword{ASOS,}
\keyword{BOR,}
\keyword{COCORAHS,UPPERAIR,}
\keyword{COOP,}
\keyword{COOP,UKN,}
\keyword{GHCND,}
\keyword{MPRC,}
\keyword{MSNT,}
\keyword{Mesonet,}
\keyword{NEON,}
\keyword{NRCS,}
\keyword{OTHER,}
\keyword{RNRCS}
\keyword{SCAN,}
\keyword{SNOW,}
\keyword{SNTL,}
\keyword{SNTLT,}
\keyword{USCRN,USRCRN,AWOS,}
\keyword{USGS,}
\keyword{USRCRN,}
\keyword{WBAN,}
\keyword{Z10,}
