% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uscrn_data.R
\name{getUSCRNData}
\alias{getUSCRNData}
\title{Pull data from meterologic stations in the USCRN network}
\usage{
getUSCRNData(timeScale, stationID, timeBgn, timeEnd)
}
\arguments{
\item{timeScale}{The time scale of measurements to return.
Enter 'monthly', 'daily', 'hourly', or 'subhourly'.}

\item{stationID}{The WBAN ID of a valid USCRN site.}

\item{TimeBgn}{A YYYY-MM-DD hh:mm:ss formated start time, in UTC.}

\item{TimeEnd}{A YYYY-MM-DD hh:mm:ss formated end time, in UTC.}
}
\value{
Data frame of USCRN data.
}
\description{
Provide a valid GHCND or WBAN ID for a site in the USCRN, the desired
timescale of measurements, and requested begin and end times in POSIX format,
and a data frame will be returned to the global environment.
}
\examples{
\dontrun{
timeScale <- "subhourly"
stationID <- "USW00003047"
timeBgn <- as.POSIXct("2014-04-01 00:00:01", format="\%Y-\%m-\%d \%H:\%M:\%S", tz="UTC")
timeEnd <- as.POSIXct("2015-02-01 00:00:00", format = "\%Y-\%m-\%d \%H:\%M:\%S", tz="UTC")

grabUSCRN(timeScale = timeScale, TimeBgn = TimeBgn, TimeEnd = TimeEnd, stationID = stationID)
}
}
\seealso{
Currently none
}
\author{
Robert Lee \email{rlee@battelleecology.org}\cr
}
\keyword{USCRN,}
\keyword{commissioning}
\keyword{data}
\keyword{data,}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
